fixup - refactor into component and page
